cv-service/requirements.txt
# FastAPI framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Computer Vision
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
numpy==2.2.2
pillow==11.1.0

# Machine Learning (optional, for future YOLOv8)
# torch==2.5.1
# torchvision==0.20.1
# ultralytics==8.3.66

# Utilities
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.29.1

# Image processing
scikit-image==0.25.0
